+++ app/src/main/java/ee/ria/DigiDoc/android/main/settings/access/TsaUrlPreferenceDialogFragment.java
@@ -1,25 +1,7 @@
-/*
- * app
- * Copyright 2017 - 2023 Riigi Infosüsteemi Amet
- *
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 2.1 of the License, or (at your option) any later version.
- *
- * This library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this library; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
- *
- */
+package ee.ria.DigiDoc.android.main.settings.access;
 
-package ee.ria.DigiDoc.android.main.settings;
 
+import static android.view.ViewGroup.LayoutParams.WRAP_CONTENT;
 import static android.view.accessibility.AccessibilityEvent.TYPE_ANNOUNCEMENT;
 
 import android.app.Dialog;
@@ -34,36 +16,41 @@ import androidx.appcompat.widget.AppCompatEditText;
 import androidx.preference.EditTextPreferenceDialogFragmentCompat;
 
 import ee.ria.DigiDoc.R;
+import ee.ria.DigiDoc.android.Application;
 import ee.ria.DigiDoc.android.accessibility.AccessibilityUtils;
 import ee.ria.DigiDoc.android.utils.SecureUtil;
 import ee.ria.DigiDoc.android.utils.TextUtil;
 import ee.ria.DigiDoc.android.utils.display.DisplayUtil;
+import ee.ria.DigiDoc.configuration.ConfigurationProvider;
 
-public class UUIDPreferenceDialogFragment extends EditTextPreferenceDialogFragmentCompat {
+public class TsaUrlPreferenceDialogFragment extends EditTextPreferenceDialogFragmentCompat {
 
     @Override
     protected void onBindDialogView(View view) {
         super.onBindDialogView(view);
-        UUIDPreference uuidPreference = getUUIDPreference();
-        if (uuidPreference != null) {
-            CheckBox checkBox = uuidPreference.getCheckBox();
+        TsaUrlPreference tsaUrlPreference = getTsaUrlPreference();
+        if (tsaUrlPreference != null) {
+            ConfigurationProvider configurationProvider = ((Application) getContext().getApplicationContext()).getConfigurationProvider();
+            CheckBox checkBox = tsaUrlPreference.getCheckBox();
 
             AppCompatEditText appCompatEditText = TextUtil.getTextView(view);
 
-            uuidPreference.setOnBindEditTextListener(editText -> {
+            tsaUrlPreference.setOnBindEditTextListener(editText -> {
                 checkBox.setChecked(false);
-                editText.setText(uuidPreference.getText());
+                editText.setText(tsaUrlPreference.getText());
             });
 
             if (appCompatEditText != null) {
                 checkBox.setOnCheckedChangeListener((buttonView, isChecked) -> {
                     appCompatEditText.setEnabled(!isChecked);
                     if (isChecked) {
-                        disableTextViewOnChecked(appCompatEditText);
+                        disableTextViewOnChecked(appCompatEditText, configurationProvider);
                     }
                 });
 
-                checkBox.setChecked(TextUtils.isEmpty(uuidPreference.getText()));
+                checkBox.setChecked(TextUtils.isEmpty(tsaUrlPreference.getText()));
+
+                SettingsAccessView.setTsaCertificateViewVisibleValue(!checkBox.isChecked());
 
                 ViewGroup parent = ((ViewGroup) appCompatEditText.getParent());
                 View oldCheckBox = appCompatEditText.findViewById(checkBox.getId());
@@ -76,15 +63,16 @@ public class UUIDPreferenceDialogFragment extends EditTextPreferenceDialogFragme
                         ((ViewGroup) oldParent).removeView(checkBox);
                     }
                     parent.addView(checkBox, ViewGroup.LayoutParams.MATCH_PARENT,
-                            ViewGroup.LayoutParams.WRAP_CONTENT);
+                            WRAP_CONTENT);
                 }
             }
         }
     }
 
-    private void disableTextViewOnChecked(AppCompatEditText appCompatEditText) {
+    private void disableTextViewOnChecked(AppCompatEditText appCompatEditText,
+                                          ConfigurationProvider configurationProvider) {
         appCompatEditText.setText(null);
-        appCompatEditText.setHint("00000000-0000-0000-0000-000000000000");
+        appCompatEditText.setHint(configurationProvider.getTsaUrl());
         appCompatEditText.clearFocus();
     }
 
@@ -95,8 +83,8 @@ public class UUIDPreferenceDialogFragment extends EditTextPreferenceDialogFragme
         return dialog;
     }
 
-    private UUIDPreference getUUIDPreference() {
-        return (UUIDPreference) this.getPreference();
+    private TsaUrlPreference getTsaUrlPreference() {
+        return (TsaUrlPreference) this.getPreference();
     }
 
     @Override
