+++ app/src/main/res/drawable/ic_icon_access_padded.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+    <!-- Transparent drawable for padding -->
+    <item android:drawable="@android:color/transparent" android:left="8dp" android:right="8dp" />
+
+    <!-- Your actual drawable icon -->
+    <item android:drawable="@drawable/ic_icon_diagnostics" android:left="8dp" android:right="8dp" />
+</layer-list>
