+++ app/src/main/res/layout/main_settings_access.xml
@@ -1,0 +1,174 @@
+<?xml version="1.0" encoding="utf-8"?>
+<merge xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:parentTag="androidx.coordinatorlayout.widget.CoordinatorLayout"
+    tools:layout_width="match_parent"
+    tools:layout_height="match_parent">
+
+    <com.google.android.material.appbar.AppBarLayout
+        android:id="@id/appBar"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
+
+        <android.widget.Toolbar
+            android:id="@id/toolbar"
+            android:theme="@style/ThemeOverlay.Application.ActionBar"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:accessibilityPaneTitle="@string/main_settings_access_button"
+            app:navigationIcon="?attr/homeAsUpIndicator"
+            app:navigationContentDescription="@string/back"
+            app:title="@string/main_settings_access_button" />
+
+    </com.google.android.material.appbar.AppBarLayout>
+
+    <ScrollView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:paddingStart="@dimen/material_content_edge_margin_horizontal"
+        android:paddingEnd="@dimen/material_content_edge_margin_horizontal"
+        android:orientation="vertical"
+        android:gravity="center_vertical"
+        app:layout_behavior="@string/appbar_scrolling_view_behavior">
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:orientation="vertical">
+
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:orientation="vertical">
+
+                <androidx.fragment.app.FragmentContainerView
+                    android:id="@id/mainSettingsFragment"
+                    android:name="ee.ria.DigiDoc.android.main.settings.access.SettingsFragment"
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent" />
+            </LinearLayout>
+
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:paddingTop="18dp"
+                android:paddingStart="16dp"
+                android:paddingEnd="8dp"
+                android:orientation="horizontal">
+
+                <TextView
+                    style="@style/MaterialTypography.Regular.Body1"
+                    android:textSize="16sp"
+                    android:layout_width="0dp"
+                    android:layout_height="wrap_content"
+                    android:text="@string/main_settings_open_all_filetypes_title"
+                    android:layout_weight="2">
+                </TextView>
+
+                <androidx.appcompat.widget.SwitchCompat
+                    android:id="@+id/mainSettingsOpenAllFileTypes"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:paddingStart="0dp"
+                    android:paddingEnd="0dp"
+                    android:minWidth="0dp"
+                    android:minHeight="0dp"
+                    android:layout_gravity="end">
+                </androidx.appcompat.widget.SwitchCompat>
+            </LinearLayout>
+
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:paddingTop="18dp"
+                android:paddingStart="16dp"
+                android:paddingEnd="8dp"
+                android:orientation="horizontal">
+
+                <TextView
+                    style="@style/MaterialTypography.Regular.Body1"
+                    android:textSize="16sp"
+                    android:layout_width="0dp"
+                    android:layout_height="wrap_content"
+                    android:text="@string/main_settings_allow_screenshots_title"
+                    android:layout_weight="2">
+                </TextView>
+
+                <androidx.appcompat.widget.SwitchCompat
+                    android:id="@+id/mainSettingsAllowScreenshots"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:paddingStart="0dp"
+                    android:paddingEnd="0dp"
+                    android:minWidth="0dp"
+                    android:minHeight="0dp"
+                    android:layout_gravity="end">
+                </androidx.appcompat.widget.SwitchCompat>
+            </LinearLayout>
+
+            <LinearLayout
+                android:id="@id/mainSettingsTsaCertificateContainer"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:paddingTop="18dp"
+                android:paddingStart="16dp"
+                android:paddingEnd="8dp"
+                android:orientation="vertical">
+
+                <TextView
+                    style="@style/MaterialTypography.Regular.Body1"
+                    android:textSize="16sp"
+                    android:layout_height="wrap_content"
+                    android:layout_width="match_parent"
+                    android:paddingTop="4dp"
+                    android:paddingBottom="4dp"
+                    android:text="@string/main_settings_timestamp_cert_title"
+                    android:layout_weight="2">
+                </TextView>
+
+                <TextView
+                    android:id="@id/mainSettingsTsaCertificateIssuedTo"
+                    style="@style/MaterialTypography.Regular.Body1"
+                    android:textSize="16sp"
+                    android:layout_height="wrap_content"
+                    android:layout_width="match_parent"
+                    android:paddingTop="4dp"
+                    android:paddingBottom="4dp"
+                    android:text="@string/main_settings_timestamp_cert_issued_to_title"
+                    android:layout_weight="2">
+                </TextView>
+
+                <TextView
+                    android:id="@id/mainSettingsTsaCertificateValidTo"
+                    style="@style/MaterialTypography.Regular.Body1"
+                    android:textSize="16sp"
+                    android:layout_height="wrap_content"
+                    android:layout_width="match_parent"
+                    android:paddingTop="4dp"
+                    android:paddingBottom="4dp"
+                    android:text="@string/main_settings_timestamp_cert_valid_to_title"
+                    android:layout_weight="2">
+                </TextView>
+
+                <Button
+                    android:id="@id/mainSettingsTsaCertificateAddCertificateButton"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_gravity="center"
+                    android:text="@string/main_settings_timestamp_cert_add_certificate_button" />
+
+                <Button
+                    android:id="@id/mainSettingsTsaCertificateShowCertificateButton"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_gravity="center"
+                    android:text="@string/main_settings_timestamp_cert_show_certificate_button" />
+
+            </LinearLayout>
+
+        </LinearLayout>
+
+    </ScrollView>
+
+</merge>
