+++ app/src/main/res/values/ids.xml
@@ -65,7 +65,12 @@
 
     <!-- Main Settings -->
     <item type="id" name="mainSettingsScreen" />
+    <item type="id" name="mainSettingsAccessScreen" />
     <item type="id" name="mainSettingsFragment" />
+    <item type="id" name="mainSettingsAccessCategory" />
+    <item type="id" name="mainSettingsRoleAndAddressCategory" />
+    <item type="id" name="mainSettingsProxyCategory" />
+    <item type="id" name="mainSettingsSSLCertificateCategory" />
     <item type="id" name="mainSettingsOpenAllFileTypes" />
     <item type="id" name="mainSettingsAllowScreenshots" />
     <item type="id" name="mainSettingsAccessToSigningService" />
